Spark/backend/requirements.txt
flask
flask_pymongo
python-dotenv
pandas
numpy
haversine
scikit-learn
ortools 